% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pubmlst.R
\name{download_latest_sequences}
\alias{download_latest_sequences}
\title{Download the latest PubMLST Campylobacter jejuni and coli sequences.}
\usage{

  download_latest_sequences(url = "http://pubmlst.org/data/profiles/campylobacter.txt")
}
\arguments{
\item{url}{the PubMLST URL to use.}
}
\value{
A data frame of sequence types, their allelic profile, and clonal complex.
}
\description{
Download the latest PubMLST Campylobacter jejuni and coli sequences.
}
\examples{
mlst <- download_latest_sequences()
head(mlst)
}
\seealso{
\url{http://pubmlst.org}
}

