% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pubmlst.R
\name{find_matching_profiles}
\alias{find_matching_profiles}
\title{Find the matching profiles in the PubMLST Campylobacter jejuni and coli database.}
\usage{
find_matching_profiles(profile, pubmlst)
}
\arguments{
\item{profile}{the allelic profile}

\item{pubmlst}{the database to match against}
}
\value{
A boolean vector of which rows in pubmlst match
}
\description{
Imputes missing alleles as appropriate.
}
\examples{
mlst    <- download_latest_sequences()
matches <- find_matching_profiles(c(1, 2, 1, 54, NA, 4, 1, 5), mlst)
head(mlst)
}

