% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmlst.R
\name{impute_mlst_in_data}
\alias{impute_mlst_in_data}
\title{Impute MLST allelic profiles by matching profiles in the PubMLST Campylobacter jejuni and coli database.}
\usage{
impute_mlst_in_data(db, cols_mlst = c("ASP", "GLN", "GLT", "GLY", "PGM",
  "TKT", "UNC"))
}
\arguments{
\item{db}{the database or data.frame.}

\item{cols_mlst}{the columns in the data representing the seven loci.}
}
\value{
db, with additional columns ST, CC, and Coli, where cols_mlst have been imputed.
}
\description{
Imputes missing alleles if there is a unique match in PubMLST, and adds sequence type (ST), clonal complex (CC)
and c.coli (Coli, logical) columns to the given database or data.frame.
}
\details{
Alleles marked as "NEW" are interpreted as new alleles, and given a large allele number.
}
\examples{
df <- data.frame(id=c("A", "B"), ASP=c(2,2), GLN=c(1,4), GLT=c(54,1), 
                 GLY=c(NA, 2), PGM=c(4,2), TKT=c(1,1), UNC=c(5,NA))
impute_mlst_in_data(df)
}

