% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmlst.R
\name{impute_sequences}
\alias{impute_sequences}
\title{Impute allelic profiles by matching profiles in the PubMLST Campylobacter jejuni and coli database.}
\usage{
impute_sequences(mlst)
}
\arguments{
\item{mlst}{the allelic profiles to match as a numeric matrix or data.frame of 7 columns.}
}
\value{
A data frame of 9 columns, representing the imputed profile, the ST and the number of imputed alleles.
}
\description{
Imputes missing alleles if there is a unique match in PubMLST.
}
\examples{
profiles <- matrix(c(2, 1, 54, NA, 4, 1,  5,
                     2, 4,  1,  2, 2, 1, NA), nrow=2, byrow=TRUE)
impute_sequences(profiles)
}
