% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmlst.R
\name{find_matching_profiles}
\alias{find_matching_profiles}
\title{Find the matching profiles in the PubMLST Campylobacter jejuni and coli database.}
\usage{
find_matching_profiles(profile)
}
\arguments{
\item{profile}{the allelic profile}
}
\value{
A boolean vector of which rows in pubmlst match
}
\description{
Profiles with missing alleles return all matching profiles.
}
\examples{
matches <- find_matching_profiles(c(2, 1, 54, NA, 4, 1, 5))
pubmlst[matches,]
}

